+++ ui/packages/shared/profile/src/ProfileIcicleGraph/IcicleGraphArrow/index.tsx
@@ -244,9 +244,6 @@ export const IcicleGraphArrow = memo(function IcicleGraphArrow({
   const highlightSimilarStacksSetName = useMemo(() => {
     return highlightSimilarStacksPreference ? setHoveringName : noop;
   }, [highlightSimilarStacksPreference]);
-  const highlightSimilarStacksSetLevel = useMemo(() => {
-    return highlightSimilarStacksPreference ? setHoveringLevel : noop;
-  }, [highlightSimilarStacksPreference]);
   const highlightSimilarStacksRow = highlightSimilarStacksPreference ? hoveringRow : null;
   const path = useMemo(() => {
     return [];
@@ -283,7 +280,7 @@ export const IcicleGraphArrow = memo(function IcicleGraphArrow({
               isRoot={true}
               searchString={(currentSearchString as string) ?? ''}
               setHoveringRow={setHoveringRow}
-              setHoveringLevel={highlightSimilarStacksSetLevel}
+              setHoveringLevel={setHoveringLevel}
               sortBy={sortBy}
               darkMode={isDarkMode}
               compareMode={compareMode}
@@ -321,7 +318,6 @@ export const IcicleGraphArrow = memo(function IcicleGraphArrow({
     colorForSimilarNodes,
     highlightSimilarStacksPreference,
     path,
-    highlightSimilarStacksSetLevel,
     highlightSimilarStacksSetName,
   ]);