+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.762](https://github.com/parca-dev/parca/compare/@parca/web@0.16.761...@parca/web@0.16.762) (2024-10-19)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.761](https://github.com/parca-dev/parca/compare/@parca/web@0.16.760...@parca/web@0.16.761) (2024-10-15)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.760](https://github.com/parca-dev/parca/compare/@parca/web@0.16.759...@parca/web@0.16.760) (2024-10-14)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.759](https://github.com/parca-dev/parca/compare/@parca/web@0.16.758...@parca/web@0.16.759) (2024-10-06)
 
 **Note:** Version bump only for package @parca/web